refactor: unifies tests into a single file
